+++ consensus/pending_votes/pending_votes_test.go
@@ -1,4 +1,4 @@
-package consensus
+package pending_votes
 
 import (
 	"testing"
